requirements.txt
pika==1.3.2
pre_commit==4.2.0
psycopg==3.2.9
psycopg_binary==3.2.9
pydantic==2.11.7
pydantic-settings==2.10.1
pytest==8.4.1
pytest-playwright==0.7.0
python-dotenv==1.1.1
slackclient==2.9.4
fake-useragent==2.2.0
requests~=2.32.4
